Revert "Merge branch 'main' into Libre311-AttributeFrontend"
This reverts commit 884f717af8c5d481aed6d5e4aebb599127384b85, reversing
changes made to e192386c4b9fc1b55cc8d4efc2fbbed5ca657595.
