+++ app/src/main/java/app/model/servicerequest/ServiceRequestStatus.java
@@ -17,7 +17,7 @@ package app.model.servicerequest;
 import com.fasterxml.jackson.annotation.JsonValue;
 
 public enum ServiceRequestStatus {
-    OPEN, IN_PROGRESS, CLOSED;
+    OPEN, CLOSED;
 
     @Override
     @JsonValue
